requirements.txt
huggingface_hub
streamlit
pillow
transformers
pycountry